Vistas

+++ src/main/java/com/iudigital/floristeriaapi/repositories/IFloresRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.iudigital.floristeriaapi.models.FloresModel;
 
 @Repository
-public interface IFloresRepository extends JpaRepository<FloresModel, Integer>{
+public interface IFloresRepository extends JpaRepository<FloresModel, Long>{
 
   
 }
